# openAPI.yaml
openapi: 3.0.0
info:
  title: Water Quality Prediction API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid data format or user already exists
        '500':
          description: An unexpected error occurred
  /login:
    post:
      summary: Login as an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in successfully
        '400':
          description: Invalid credentials or data format
        '500':
          description: An unexpected error occurred
  /send_readings:
    post:
      summary: Save sensor readings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ph:
                  type: number
                turbidity:
                  type: number
                conductivity:
                  type: number
                sensor_id:
                  type: string
              required:
                - ph
                - turbidity
                - conductivity
                - sensor_id
      responses:
        '200':
          description: Readings Saved
        '400':
          description: Invalid input
        '500':
          description: Unexpected Error Occurred
  /get_readings:
    get:
      summary: Get sensor readings
      parameters:
        - in: query
          name: sensor_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Returned sensor readings
          content:
            application/json:
              schema:
                type: object
                properties:
                  readings:
                    type: object
        '404':
          description: No readings found
        '500':
          description: Unexpected Error Occurred
  /add_sensor:
    post:
      summary: Add a new sensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sensor_id:
                  type: string
              required:
                - sensor_id
      responses:
        '200':
          description: Sensor Added
        '500':
          description: Unexpected Error Occurred
  /predict:
    post:
      summary: Get water quality prediction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    turbidity:
                      type: number
                    ph:
                      type: number
                    conductivity:
                      type: number
      responses:
        '200':
          description: Successful prediction
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    type: number
                  output:
                    type: number
                  all_output:
                    type: array
                    items:
                      type: number
        '400':
          description: Invalid input
        '500':
          description: Unexpected Error Occurred